Add exception handling for post collisions
Very rarely, it is possible that two people can post at the same time in
such a way that one will make it through and the other will not see that
the latest post ID in the boards table has updated, resulting in a
trigger throwing an error: "That post already exists!"

Exception handling has been added for this. Again, it's pretty rare that
his happens - I did it from localhost using two executions of curl
running in parallel, and it still only happened about once in 15 tries.
Still, it's better to print something than nothing at all in that one
odd time that it does.